Add getListData into own component and implement component for displayed data
